ToolboxAI-Roblox-Environment/requirements.txt
# ToolboxAI Roblox Environment - Complete Requirements
# Python 3.11+ required

# Core Dependencies
fastapi==0.116.1
uvicorn[standard]==0.35.0
flask==3.0.0
flask-cors==6.0.1

# LangChain & AI
langchain==0.3.27
langchain-openai==0.3.32
langchain-community==0.3.29
langgraph==0.6.6
openai==1.106.0
tiktoken==0.11.0

# Database & Storage
sqlalchemy==2.0.23
alembic==1.13.0
asyncpg==0.30.0
redis==6.4.0
motor==3.3.2

# Authentication & Security
python-jose[cryptography]==3.4.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.20
requests-oauthlib==2.0.0

# Data Validation & Serialization
pydantic==2.11.6
pydantic-settings==2.10.1
python-dotenv==1.0.0
pyyaml==6.0.1

# WebSocket & Real-time
websockets==15.0.1
python-socketio==5.10.0

# HTTP & API Clients
httpx==0.28.1
aiohttp==3.9.2
requests==2.32.5

# Monitoring & Logging
prometheus-client==0.22.1
structlog==25.4.0
colorama==0.4.6
sentry-sdk[fastapi,sqlalchemy,pure_eval]==2.37.1

# Testing
pytest==8.4.2
pytest-asyncio==1.1.0
pytest-cov==6.2.1
pytest-mock==3.12.0

# Development Tools
black==25.1.0
flake8==7.3.0
mypy==1.7.1
ipython==9.5.0

# System Monitoring
psutil==7.0.0
py-cpuinfo==9.0.0

# Vector Embeddings (for MCP)
numpy==2.3.2
sentence-transformers==5.1.0
faiss-cpu==1.8.0

# Educational Content Tools
wikipedia-api==0.8.1
duckduckgo-search==8.1.1

# Utilities
python-dateutil==2.9.0.post0
pytz==2025.2
aiofiles==24.1.0
tenacity==9.1.2

# Optional: Production Server
gunicorn==23.0.0
supervisor==4.2.5

# Optional: Documentation
mkdocs==1.5.3
mkdocs-material==9.5.2

# ToolboxAI Utils Dependencies
psutil==7.0.0
urllib3==2.2.3
